Refactor INSZ validation to use int for birth year and improve date handling logic

+++ src/dotnet/johnverbiest.belgian.insz/InszNumber.cs
@@ -7,7 +7,7 @@ public record InszNumber
 {
     private readonly Lazy<string> _inszString;
     private readonly Lazy<DateTime?> _birthDate;
-    private readonly Lazy<short?> _birthYear;
+    private readonly Lazy<int?> _birthYear;
     private readonly Lazy<string> _formattedString;
 
     public InszNumber()
@@ -15,7 +15,7 @@ public record InszNumber
         _inszString = new Lazy<string>(() => InszValidator.GetInszString(Value), isThreadSafe: true);
         _birthDate = new Lazy<DateTime?>(() => InszValidator.GetDate(_inszString.Value), isThreadSafe: true);
         _formattedString = new Lazy<string>(() => $"{_inszString.Value.Substring(0, 2)}.{_inszString.Value.Substring(2, 2)}.{_inszString.Value.Substring(4, 2)}-{_inszString.Value.Substring(6, 3)}.{_inszString.Value.Substring(9, 2)}", isThreadSafe:true);
-        _birthYear = new Lazy<short?>(() => InszValidator.GetBirthYear(_inszString.Value), isThreadSafe: true);
+        _birthYear = new Lazy<int?>(() => InszValidator.GetBirthYear(_inszString.Value), isThreadSafe: true);
     }
 
     /// <summary>
@@ -39,6 +39,6 @@ public record InszNumber
 
     public DateTime? BirthDate => IsValid == true ? InszValidator.GetDate(_inszString.Value) : null;
     
-    public short? BirthYear => IsValid == true ? _birthYear.Value : null;
+    public int? BirthYear => IsValid == true ? _birthYear.Value : null;
 }
